+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 - `RDF.Description.take/2` creates a description from another one by limiting 
   its statements to a set of predicates
+- `RDF.Graph.take/3` creates a graph from another one by limiting 
+  its statements to a set of subjects and optionally also a set of predicates
 - Mix formatter configuration for using `defvocab` without parens